fix: update toUserProfileResponse

+++ src/main/java/com/DevLink/backend/dto/UserProfileResponse.java
@@ -13,7 +13,7 @@ public record UserProfileResponse(
         String bio,
         String avatar,
         LocalDateTime createdAt,
-        int projectsCount,
+        long projectsCount,
         int collaborationsCount,
         String githubUrl,
         String gitlabUrl
